requirements.txt
pandas==1.2.4
numpy==1.20.1
Image==1.5.33
Pillow==8.3.1
progressbar33==2.4
pyautogui==0.9.53
pynput==1.7.3
PyQt5==5.15.4
pytesseract==0.3.8
